Update README.md
Heart disease is one of the leading causes of death worldwide. Healthcare professionals and researchers face challenges in analyzing large volumes of patient data to identify patterns, risk factors, and key indicators that contribute to heart disease. The unorganized and complex nature of medical data makes it difficult to draw actionable insights in real time, which delays preventive measures and critical decision-making.